pyproject.toml
[project]
name = "ivory-tower"
version = "0.1.0"
description = "Multi-agent deep research orchestrator"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.15",
    "rich>=13",
    "pyyaml>=6",
]

[project.optional-dependencies]
adversarial = ["gepa>=0.1.0"]

[project.scripts]
ivory = "ivory_tower.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ivory_tower"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-m 'not live'"
markers = [
    "live: real e2e tests that call external agents and APIs (run with: pytest -m live)",
]

[dependency-groups]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "gepa>=0.1.0",
]
